style: rename base scraper
